+++ source/commands/moderation/index.js
@@ -1,4 +1,4 @@
-const { PermissionFlagsBits } = require('discord.js');
+const { PermissionFlagsBits, InteractionContextType } = require('discord.js');
 const { CommandWrapper } = require('../../classes');
 const { createSubcommandMappings } = require('../../util/fileUtil.js');
 
@@ -10,7 +10,7 @@ const { slashData: subcommandSlashData, executeDictionary: subcommandExecuteDict
 	"xppenalty.js",
 	"bountybotban.js"
 ]);
-module.exports = new CommandWrapper(mainId, "BountyBot moderation tools", PermissionFlagsBits.ManageRoles, false, false, 3000,
+module.exports = new CommandWrapper(mainId, "BountyBot moderation tools", PermissionFlagsBits.ManageRoles, false, [InteractionContextType.Guild], 3000,
 	(interaction, database, runMode) => {
 		subcommandExecuteDictionary[interaction.options.getSubcommand()](interaction, database, runMode);
 	}