+++ source/commands/scoreboard.js
@@ -1,8 +1,9 @@
+const { InteractionContextType } = require('discord.js');
 const { CommandWrapper } = require('../classes');
 const { buildSeasonalScoreboardEmbed, buildOverallScoreboardEmbed } = require('../util/embedUtil');
 
 const mainId = "scoreboard";
-module.exports = new CommandWrapper(mainId, "View the XP scoreboard", null, false, false, 3000,
+module.exports = new CommandWrapper(mainId, "View the XP scoreboard", null, false, [InteractionContextType.BotDM, InteractionContextType.Guild, InteractionContextType.PrivateChannel], 3000,
 	/** View the XP scoreboard */
 	(interaction, database, runMode) => {
 		if (interaction.options.getString("scoreboard-type") === "season") {